requirements.txt
asgiref==3.2.10
Django==3.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
